requirements.txt
--index-url https://pypi.devinfra.sentry.io/simple

black==22.6.0
blinker==1.5
click==8.1.3
clickhouse-driver==0.2.4
confluent-kafka==2.1.1
datadog==0.21.0
flake8==5.0.4
Flask==2.2.2
honcho==1.1.0
python-jose[cryptography]==3.3.0
jsonschema==4.16.0
fastjsonschema==2.16.2
packaging==21.3
# Do not update parsimonious, there is a performance issue in newer versions
parsimonious==0.8.1
progressbar2==4.0.0
pytest==7.1.3
pytest-cov==3.0.0
pytest-watch==4.2.0
python-dateutil==2.8.2
python-rapidjson==1.8
pytz==2022.2.1
redis==4.3.4
sentry-arroyo==2.14.1
sentry-kafka-schemas==0.1.21
sentry-redis-tools==0.1.7
sentry-relay==0.8.27
sentry-sdk==1.28.0
simplejson==3.17.6
structlog==22.3.0
structlog-sentry==2.0.0
sql-metadata==2.6.0
typing-extensions==4.5.0
urllib3==1.26.12
pyuwsgi==2.0.20
Werkzeug==2.2.3
PyYAML==6.0
sqlparse==0.4.2
google-api-python-client==2.88.0
